split Siglip && Adjust doc
